% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{as_TransitionMatrix}
\alias{as_TransitionMatrix}
\title{Transition Matrix}
\usage{
as_TransitionMatrix(tmat, state_labels = colnames(tmat))
}
\arguments{
\item{tmat}{a k-by-k integer-ish matrix where \code{tmat[i,j]} is missing if the
transition is impossible and an integer otherwise. The integers must be
unique and sequential since they are used to index into the list of
transition distributions.}

\item{state_labels}{character vector of length k with labels for the states;
populated from the column names of \code{tmat} if available.}
}
\value{
A matrix of class \code{TransitionMatrix}.
}
\description{
Check a multi-state transition matrix for consistency and
turn it into a matrix of class \code{TransitionMatrix}
}
