% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calibration_probability}
\alias{calibration_probability}
\title{Transition probability for approximate likelihood inference}
\usage{
calibration_probability(from, to, probability, standard_error)
}
\arguments{
\item{from}{character of length one, starting state}

\item{to}{character, target state}

\item{probability}{numeric vector of transition probability}

\item{standard_error}{numeric vector of standard errors for transition
probabilities}
}
\description{
Transition probability for approximate likelihood inference
}
