% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calibration_survival_curve}
\alias{calibration_survival_curve}
\title{Survival curve for approximate likelihood inference}
\usage{
calibration_survival_curve(from, to, time, survival, standard_error)
}
\arguments{
\item{from}{character of length one, starting state}

\item{to}{character vector of events in composite time-to-event endpoint}

\item{time}{numeric vector of times}

\item{survival}{numeric vector of survival proportions at \code{time}}

\item{standard_error}{numeric vector of standard errors for survival
proportions}
}
\description{
Survival curve for approximate likelihood inference
}
