% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{censor}
\alias{censor}
\title{Censor multi-state data set}
\usage{
censor(data, t = Inf)
}
\arguments{
\item{data}{multi-state data set}

\item{t}{positive numerical, global maximal censoring time point}
}
\description{
Restrict multi-state data set to given time point.
}
