% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmsimr-package.R
\docType{data}
\name{example_calibration_data}
\alias{example_calibration_data}
\title{This is data to be included in my package}
\format{
An object of class \code{list} of length 4.
}
\usage{
example_calibration_data
}
\description{
This is data to be included in my package
}
\keyword{datasets}
