% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Exponential}
\alias{Exponential}
\title{Exponential Distribution}
\usage{
Exponential(lambda, optimize = FALSE)
}
\arguments{
\item{lambda}{rate}
}
\value{
A \href{survival-distribution}{survival distribution} of class \code{Exponential}.
}
\description{
description
}
\seealso{
\link{survival-distribution}
}
\concept{survival distributions}
