% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient descent}
\usage{
gradient_descent(fn, gr, init, rate, niter)
}
\arguments{
\item{fn}{function to optimize}

\item{gr}{gradient}

\item{init}{initial parameter variables}

\item{rate}{learning rate}

\item{niter}{number of iteration}
}
\description{
Gradient descent
}
