% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{kaplan_meier}
\alias{kaplan_meier}
\title{Kaplan-Meier estimator of time-to-first-event}
\usage{
kaplan_meier(msm, from, to, data = simulate(msm), tidy = TRUE)
}
\arguments{
\item{msm}{multi-state model object of class \code{\link{MSM}}}

\item{from}{character of length one, starting state}

\item{to}{character vector of events in composite time-to-event endpoint}

\item{data}{multi-state data set}

\item{tidy}{logical, if true data is returned as \verb{[tibble::tibble]}}
}
\description{
...
}
