% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Sample-based log-likelihood for summary data}
\usage{
log_likelihood(
  msm,
  ...,
  nsim = attr(msm, "m"),
  tmax = get_tmax(msm),
  grid_size = attr(msm, "grid_size"),
  hazard_ratios = rep(1, length(msm)),
  seed = NULL
)
}
\arguments{
\item{msm}{multi-state model object of class \code{\link{MSM}}}

\item{...}{a list of calibration objects TODO: explain}

\item{nsim}{integer, number of iterations for simulation}

\item{tmax}{numeric, maximal time when simulating from model}

\item{grid_size}{integer, number of grid points to use when calculating the
cumulative hazard grid for simulation}

\item{hazard_ratios}{length(states)-numeric with hazard ratios to modify
transition hazards with}

\item{seed}{integer, random seed to use}
}
\description{
...
}
