% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSM.R
\name{MSM}
\alias{MSM}
\alias{get_tmat}
\alias{get_tmax}
\alias{get_state_labels}
\alias{get_transition_distribution}
\alias{update_parameters}
\title{Multi-State Model}
\usage{
MSM(..., tmat, tmax = NULL, m = 2^11, grid_size = 500)

get_tmat(msm)

get_tmax(msm)

get_state_labels(msm)

get_transition_distribution(msm, index)

update_parameters(msm, params)
}
\arguments{
\item{...}{list of \link[=survival-distribution]{survival distributions};
distributions are mapped to transitions using the indices specified
in the transition matrix \code{tmat}}

\item{tmat}{transition matrix for the multi-state model}

\item{tmax}{maximal time when simulating from model}

\item{msm}{a \code{MSM} multi-state object}

\item{index}{the integer index of the distribution to return}

\item{params}{numeric vector with new paramter values}
}
\description{
...

\code{get_tmat()} returns the transition matrix of an \code{MSM} object.

\code{get_tmax()} returns the maximal simulation horizon of a
\code{MSM} object.

\code{get_state_labels()} ...

\code{get_transition_distribution()} returns a single distribution

update model parameters
}
\seealso{
\code{\link[=as_TransitionMatrix]{as_TransitionMatrix()}} \link{survival-distribution}
}
