% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_state_probabilities.R
\name{next_state_probabilities}
\alias{next_state_probabilities}
\title{Compute sample-based next-state probabilities}
\usage{
next_state_probabilities(msm, data = simulate(msm), as_matrix = FALSE)
}
\arguments{
\item{msm}{multi-state model object of class \code{\link{MSM}}}

\item{data}{multi-state data set}

\item{as_matrix}{logical, return as matrix?}
}
\description{
...
}
