% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\title{Time-to-first-event survival function of multi-state model}
\usage{
survival(msm, t, from, to, data = simulate(msm))
}
\arguments{
\item{msm}{multi-state model object of class \code{\link{MSM}}}

\item{t}{numeric, time points to evaluate survival function on}

\item{from}{character of length one, starting state}

\item{to}{character vector of events in composite time-to-event endpoint}

\item{data}{multi-state data set}
}
\description{
...
}
